+++ css/parser/src/parser/mod.rs
@@ -3,6 +3,9 @@ use swc_atoms::js_word;
 use swc_common::Span;
 use swc_css_ast::*;
 
+mod at_rule;
+mod style_rule;
+
 pub type PResult<T> = Result<T, Error>;
 
 pub struct Parser<'i> {
@@ -34,40 +37,6 @@ impl Parser<'_> {
             _ => self.parse_style_rule().map(From::from),
         }
     }
-
-    fn parse_at_rule(&mut self) -> PResult<AtRule> {
-        assert_eq!(self.i.cur(), Some(Token::At));
-        self.i.bump(); // '@'
-
-        let name = self.pares_word()?;
-
-        match name.sym {
-            js_word!("charset") => self.parse_at_rule().map(From::from),
-            js_word!("media") => self.parse_media_rule().map(From::from),
-            js_word!("import") => self.parse_import_rule().map(From::from),
-            js_word!("supports") => self.parse_import_rule().map(From::from),
-            js_word!("keyframes") => self.parse_keyframes_rule().map(From::from),
-            js_word!("font-face") => self.parse_font_face_rule().map(From::from),
-
-            _ => self.parse_unknown_at_rule().map(From::from),
-        }
-    }
-
-    fn parse_charset_rule(&mut self) -> PResult<CharsetRule> {}
-
-    fn parse_media_rule(&mut self) -> PResult<MediaRule> {}
-
-    fn parse_media_query(&mut self) -> PResult<MediaQuery> {}
-
-    fn parse_import_rule(&mut self) -> PResult<ImportRule> {}
-
-    fn parse_keyframes_rule(&mut self) -> PResult<KeyframesRule> {}
-
-    fn parse_font_face_rule(&mut self) -> PResult<FontFaceRule> {}
-
-    fn parse_unknown_at_rule(&mut self) -> PResult<UnknownAtRule> {}
-
-    fn parse_style_rule(&mut self) -> PResult<StyleRule> {}
 
     fn parse_property(&mut self) -> PResult<Property> {}
